+++ dotnet/demo/SmartStandards.Logging.Demo.net48/Logging/SmartStandards/MyLogTemplate.cs
@@ -1,0 +1,15 @@
+﻿using Logging.SmartStandards.TemplateHousekeeping;
+using System.ComponentModel;
+
+namespace Logging.SmartStandards {
+
+  [TypeConverter(typeof(LogEventTemplateEnumConverter))]
+  public enum MyLogTemplate {
+
+    [LogMessageTemplate("There is too much foo within bar beacause of {0}!")]
+    [LogMessageTemplate("Da ist zu viel Foo im Bar wegen {0}!", "de-de")]
+    ZuVielFooImBar = 70011
+
+  }
+
+}