+++ vers/changelog.md
@@ -4,9 +4,7 @@ This files is automatically maintained using the ['KornSW-VersioningUtil'](https
 
 ## Upcoming Changes
 
-*(none)*
-
-
+ - **new Feature**: AutoFlush (enabled by default) for TraceBusFeed
 
 ## v 2.5.0
 released **2025-05-12**, including: